Created auto-generated syntax_test
I've created a python script to auto generate the syntax tests for simple assertions. The output of the script is in the syntax_test_d3-basic_auto_generated.bp file. The script won't be included in this repo. The tests that are being auto generated have been removed from the main syntax test file (which has also been renamed).