Restructure then/else/locked/onerr blocks
Removed the start of the blocks from the specific statements that would trigger them. This will make it easier to adapt/update other MV Basic variants. The tradeoff is that invlaid statemens will be highlighted to look correct (like IF CONDITION LOCKED...). Also added an invalid scope for stray end statements.