Rework of syntax defs and tests
This rework fixes #15 (statements after labels not getting scoped). The change that the fix required was to make the statement start expression MUCH less selective. Now it's just looking for a word boundary expect for "." or "$". A side effect of this is that two successive statements (like "CRT CRT") will both be highlighted.

The refactoring was under taken to make it easier to apply updates from one syntax to another. The was accomplished by moving all of the language specific elements into variables at the top of the syntax files.